git_analysis_project/requirements.txt
# Core testing framework
pytest
pytest-cov
pytest-benchmark  # Added for performance testing

# Git interaction library (required by gitutils.py)
GitPython

# Data manipulation (required by gitutils.py)
pandas

# Visualization (required by gitutils.py)
matplotlib

# Property-based testing
hypothesis